requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0

# Data handling
h5py>=3.8.0
pyyaml>=6.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Optional: for video I/O
opencv-python>=4.7.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
